fix date

+++ app/src/main/java/com/stdio/iptesttask/domain/model/Item.kt
@@ -8,5 +8,5 @@ import kotlinx.parcelize.Parcelize
 
 @Parcelize
 @Entity
-class Item(@PrimaryKey(autoGenerate = true) var id: Int = 0, val name: String, val time: Int, val tags: String, val amount: Int) : Parcelable{
+class Item(@PrimaryKey(autoGenerate = true) var id: Int = 0, val name: String, val time: Long, val tags: String, val amount: Int) : Parcelable{
 }
